requirements_dev.txt
pytest==2.8.5
pytest-mock==0.9.0
pytest-cov==1.7.0
sphinx
sphinx_rtd_theme
gevent
lyft-stdlib==24.4.0